Class hierarchy for actions, robocopy fixes

+++ filesystem.py
@@ -6,7 +6,6 @@ from pathlib import WindowsPath
 from click import ClickException
 
 import console as con
-from folders import Folder, Library
 from robocopy import run_robocopy
 
 
@@ -50,7 +49,7 @@ def test_disk_space(dir_size_bytes, target_disk: DiskUsage) -> None:
         f"Testing free space in drive {con.style_path(target_disk.drive)}", end=""
     )
     if dir_size_bytes > target_disk.usage.free:
-        raise ClickException(f"Drive {target_disk.drive} does have enough free space")
+        raise ClickException(f"\nDrive {target_disk.drive} does have enough free space")
     con.print_success()
 
 
@@ -58,7 +57,7 @@ def test_dir_creation(path: WindowsPath) -> None:
     """Test write access by creating and deleting an empty folder"""
     con.print_(f"Testing write access to {con.style_path(path)}", end="")
     if path.exists():
-        raise ClickException(f"{path} already exists")
+        raise ClickException(f"\n{path} already exists")
     try:
         path.mkdir()
     except PermissionError as e:
@@ -80,14 +79,19 @@ def test_set_ntfs_permisisons(source: WindowsPath, target: WindowsPath) -> None:
         source.mkdir()
         run_robocopy(source, target, copy_permissions=True, quiet=True)
     except ClickException as e:
-        if "Copying NTFS Security to Destination Directory" in e.message:
+        if (
+            "Copying NTFS Security to Destination Directory" in e.message
+            or "You do not have the Manage Auditing user right" in e.message
+        ):
             raise ClickException(
                 "Unable to set permissions. Do you need to run Command Prompt as Administrator?"
             ) from e
         raise e
     finally:
-        source.rmdir()
-        target.rmdir()
+        if source.exists():
+            source.rmdir()
+        if target.exists():
+            target.rmdir()
     con.print_success()
 
 
@@ -136,11 +140,11 @@ def create_symlink(
         return
 
     if not target.exists():
-        raise ClickException(f"{target} does not exist")
+        raise ClickException(f"\n{target} does not exist")
     try:
         source.symlink_to(target, target_is_directory=True)
     except FileExistsError as e:
-        raise ClickException(f"{source} already exists") from e
+        raise ClickException(f"\n{source} already exists") from e
     except OSError as e:
         raise ClickException(
             "Permission denied when creating symlink. Run Command Prompt as Administrator or enable Windows Developer Mode."
@@ -155,7 +159,7 @@ def delete_symlink(
     if not quiet:
         con.print_(f"Deleting symlink {con.style_path(path)}", end="")
     if not path.is_symlink():
-        raise ClickException(f"{path} is not a symlink")
+        raise ClickException(f"\n{path} is not a symlink")
     if dry_run and not quiet:
         con.print_skipped()
     if dry_run:
@@ -169,7 +173,7 @@ def delete_symlink(
 def delete_folder(path: WindowsPath, dry_run: bool = False) -> None:
     con.print_(f"Deleting folder {con.style_path(path)}", end="")
     if path.is_symlink():
-        raise ClickException(f"Cannot delete. {path} is a symlink.")
+        raise ClickException(f"\nCannot delete. {path} is a symlink.")
     if dry_run:
         con.print_skipped()
         return
@@ -188,66 +192,3 @@ def delete_folder(path: WindowsPath, dry_run: bool = False) -> None:
             "[red]To tidy up, restart your PC and delete this folder manually.[/red]"
         )
     con.print_success()
-
-
-def add_folder_actions(
-    folder: Folder, library: Library, dry_run: bool, dont_copy_permissions: bool
-) -> None:
-    """
-    Filesystem actions for `add` command
-
-    Move data from `folder.source_dir` to `folder.get_target_dir()`
-    """
-    from_dir = folder.source_dir
-    to_dir = folder.get_library_subdir(library)
-    temp_dir = folder.get_temp_dir()
-    dir_size_bytes = get_dir_size(from_dir)
-    con.print_(f"Folder size: {con.style_bytes_as_gb(dir_size_bytes)}")
-
-    con.print_("\n[bold]Pre-flight checks[/bold]")
-    test_dir_creation(to_dir)
-    test_dir_creation(temp_dir)
-    # Test symlink with sibling of source dir - it should have similar permissions
-    test_symlink_creation(temp_dir, to_dir)
-    test_disk_space(dir_size_bytes, DiskUsage(to_dir.drive))
-    if not dont_copy_permissions:
-        # Use empty source directory to test permissions
-        test_set_ntfs_permisisons(temp_dir, to_dir)
-
-    con.print_("\n[bold]Actions[/bold]")
-    rename_folder(from_dir, temp_dir, dry_run=dry_run)
-    run_robocopy(
-        temp_dir,
-        to_dir,
-        dir_size_bytes,
-        dry_run=dry_run,
-        copy_permissions=not dont_copy_permissions,
-    )
-    create_symlink(from_dir, to_dir, dry_run=dry_run)
-    delete_folder(temp_dir, dry_run=dry_run)
-
-
-def remove_folder_actions(folder: Folder, library: Library, dry_run: bool) -> None:
-    """
-    Filesystem actions for `remove` command
-
-    Move data from `folder.get_library_subdir()` to `folder.source_dir`
-    """
-    from_dir = folder.get_library_subdir(library)
-    to_dir = folder.source_dir
-    temp_dir = folder.get_temp_dir()
-    dir_size_bytes = get_dir_size(from_dir)
-    con.print_(f"Folder size: {con.style_bytes_as_gb(dir_size_bytes)}")
-
-    con.print_("\n[bold]Pre-flight checks[/bold]")
-    # Sibling of source dir
-    test_dir_creation(temp_dir)
-    # Subdir of library
-    test_dir_creation(library.get_test_dir())
-    test_disk_space(dir_size_bytes, DiskUsage(to_dir.drive))
-
-    con.print_("\n[bold]Actions[/bold]")
-    run_robocopy(from_dir, temp_dir, dir_size_bytes, dry_run=dry_run)
-    delete_symlink(to_dir, dry_run=dry_run)
-    rename_folder(temp_dir, to_dir, dry_run=dry_run)
-    delete_folder(from_dir, dry_run=dry_run)
